Delete unused module in tests
And move the remaining few functions that were actually used to the
helper module.
