--- program/tests/helpers/stakepool_account.rs
@@ -1,250 +1,0 @@
-#![allow(dead_code)]
-
-use solana_program::sysvar;
-
-use super::create_token_account;
-
-use {
-    solana_program::{
-        borsh::get_packed_len, hash::Hash, program_pack::Pack, pubkey::Pubkey, system_instruction,
-        system_program,
-    },
-    solana_program_test::*,
-    solana_sdk::{
-        signature::{Keypair, Signer},
-        transaction::Transaction,
-        transport::TransportError,
-    },
-    solana_vote_program::{
-        self, vote_instruction,
-        vote_state::{VoteInit, VoteState},
-    },
-    spl_stake_pool::{self, borsh::get_instance_packed_len, stake_program, state},
-};
-
-use lido::token::Lamports;
-
-pub const TEST_STAKE_AMOUNT: u64 = 1_500_000_000;
-pub const MAX_TEST_VALIDATORS: u32 = 10_000;
-
-pub async fn create_mint(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    pool_mint: &Keypair,
-    manager: &Pubkey,
-) -> Result<(), TransportError> {
-    let rent = banks_client.get_rent().await.unwrap();
-    let mint_rent = rent.minimum_balance(spl_token::state::Mint::LEN);
-
-    let mut transaction = Transaction::new_with_payer(
-        &[
-            system_instruction::create_account(
-                &payer.pubkey(),
-                &pool_mint.pubkey(),
-                mint_rent,
-                spl_token::state::Mint::LEN as u64,
-                &spl_token::id(),
-            ),
-            spl_token::instruction::initialize_mint(
-                &spl_token::id(),
-                &pool_mint.pubkey(),
-                &manager,
-                None,
-                0,
-            )
-            .unwrap(),
-        ],
-        Some(&payer.pubkey()),
-    );
-    transaction.sign(&[payer, pool_mint], *recent_blockhash);
-    banks_client.process_transaction(transaction).await?;
-    Ok(())
-}
-
-pub async fn transfer(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    recipient: &Pubkey,
-    amount: Lamports,
-) {
-    let transaction = Transaction::new_signed_with_payer(
-        &[system_instruction::transfer(
-            &payer.pubkey(),
-            recipient,
-            amount.0,
-        )],
-        Some(&payer.pubkey()),
-        &[payer],
-        *recent_blockhash,
-    );
-    banks_client.process_transaction(transaction).await.unwrap();
-}
-
-pub async fn mint_tokens(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    mint: &Pubkey,
-    account: &Pubkey,
-    mint_authority: &Keypair,
-    amount: u64,
-) -> Result<(), TransportError> {
-    let transaction = Transaction::new_signed_with_payer(
-        &[spl_token::instruction::mint_to(
-            &spl_token::id(),
-            mint,
-            account,
-            &mint_authority.pubkey(),
-            &[],
-            amount,
-        )
-        .unwrap()],
-        Some(&payer.pubkey()),
-        &[payer, mint_authority],
-        *recent_blockhash,
-    );
-    banks_client.process_transaction(transaction).await?;
-    Ok(())
-}
-
-pub async fn get_token_balance(banks_client: &mut BanksClient, token: &Pubkey) -> u64 {
-    let token_account = banks_client.get_account(*token).await.unwrap().unwrap();
-    let account_info: spl_token::state::Account =
-        spl_token::state::Account::unpack_from_slice(token_account.data.as_slice()).unwrap();
-    account_info.amount
-}
-
-pub async fn get_token_supply(banks_client: &mut BanksClient, mint: &Pubkey) -> u64 {
-    let mint_account = banks_client.get_account(*mint).await.unwrap().unwrap();
-    let account_info =
-        spl_token::state::Mint::unpack_from_slice(mint_account.data.as_slice()).unwrap();
-    account_info.supply
-}
-
-pub async fn delegate_tokens(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    account: &Pubkey,
-    manager: &Keypair,
-    delegate: &Pubkey,
-    amount: u64,
-) {
-    let transaction = Transaction::new_signed_with_payer(
-        &[spl_token::instruction::approve(
-            &spl_token::id(),
-            &account,
-            &delegate,
-            &manager.pubkey(),
-            &[],
-            amount,
-        )
-        .unwrap()],
-        Some(&payer.pubkey()),
-        &[payer, manager],
-        *recent_blockhash,
-    );
-    banks_client.process_transaction(transaction).await.unwrap();
-}
-
-
-pub async fn create_independent_stake_account(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    stake: &Keypair,
-    authorized: &stake_program::Authorized,
-    lockup: &stake_program::Lockup,
-    stake_amount: Lamports,
-) -> Lamports {
-    let rent = banks_client.get_rent().await.unwrap();
-    let min_rent_balance =
-        Lamports(rent.minimum_balance(std::mem::size_of::<stake_program::StakeState>()));
-    let lamports = (min_rent_balance + stake_amount).unwrap();
-
-    let transaction = Transaction::new_signed_with_payer(
-        &stake_program::create_account(
-            &payer.pubkey(),
-            &stake.pubkey(),
-            authorized,
-            lockup,
-            lamports.0,
-        ),
-        Some(&payer.pubkey()),
-        &[payer, stake],
-        *recent_blockhash,
-    );
-    banks_client.process_transaction(transaction).await.unwrap();
-
-    lamports
-}
-
-pub async fn create_blank_stake_account(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    stake: &Keypair,
-) -> u64 {
-    let rent = banks_client.get_rent().await.unwrap();
-    let lamports = rent.minimum_balance(std::mem::size_of::<stake_program::StakeState>()) + 1;
-
-    let transaction = Transaction::new_signed_with_payer(
-        &[system_instruction::create_account(
-            &payer.pubkey(),
-            &stake.pubkey(),
-            lamports,
-            std::mem::size_of::<stake_program::StakeState>() as u64,
-            &stake_program::id(),
-        )],
-        Some(&payer.pubkey()),
-        &[payer, stake],
-        *recent_blockhash,
-    );
-    banks_client.process_transaction(transaction).await.unwrap();
-
-    lamports
-}
-
-pub async fn delegate_stake_account(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    stake: &Pubkey,
-    authorized: &Keypair,
-    vote: &Pubkey,
-) {
-    let mut transaction = Transaction::new_with_payer(
-        &[stake_program::delegate_stake(
-            &stake,
-            &authorized.pubkey(),
-            &vote,
-        )],
-        Some(&payer.pubkey()),
-    );
-    transaction.sign(&[payer, authorized], *recent_blockhash);
-    banks_client.process_transaction(transaction).await.unwrap();
-}
-
-pub async fn authorize_stake_account(
-    banks_client: &mut BanksClient,
-    payer: &Keypair,
-    recent_blockhash: &Hash,
-    stake: &Pubkey,
-    authorized: &Keypair,
-    new_authorized: &Pubkey,
-    stake_authorize: stake_program::StakeAuthorize,
-) {
-    let mut transaction = Transaction::new_with_payer(
-        &[stake_program::authorize(
-            &stake,
-            &authorized.pubkey(),
-            &new_authorized,
-            stake_authorize,
-        )],
-        Some(&payer.pubkey()),
-    );
-    transaction.sign(&[payer, authorized], *recent_blockhash);
-    banks_client.process_transaction(transaction).await.unwrap();
-}
