+++ program/tests/stake_deposit.rs
@@ -3,8 +3,7 @@
 mod helpers;
 
 use helpers::{
-    get_account, program_test, simple_add_validator_to_pool,
-    LidoAccounts, ValidatorAccounts
+    get_account, program_test, simple_add_validator_to_pool, LidoAccounts, ValidatorAccounts,
 };
 use lido::state::Lido;
 use lido::token::Lamports;
